dashboard-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gitpython==3.1.40
docker==6.1.3
pylint==3.0.2
python-dotenv==1.0.0
psutil==5.9.6
requests==2.31.0
python-multipart==0.0.6 